cleanup Main.js drastically
Remove all the restorable hacky stuff, just add an `getState` and `setState` property on the props of a component.
Plugin developers can use those methods to get the state or save the state from/to localStorage (or any other storage engine).

getState() and setState(state) are automatically bound per component, meaning that internally the id of the component is attached.