add a restorableComponent and enhance helper
- Everytime ance your component like restorableComponent(YourComponent) everytime you
call this.setState() (which is the same api as before) your state will automatically
be saved into the storage engine.
- Everytime your component gets mounted the getInitialState or the this.state will be extended/overriden with the stored value

- We also introduce enhance because we need an easy way to enhance components, currently we have 2 enhancers restorableComponent and withStyles(styles), this can be used as:

enhance(YourComponent, [
  restorableComponent,
  withStyles(styles)
]);

The reason withStyles is passed a styles object is because it needs one and returns a function which is compatible with the enhance interface.
Interface is:

(Component) -> newComponent